Confocal_Axon_Analysis_Pipeline/utils/Confocal/requirements.txt
# Requirements for L1CAM Analysis - Branch-based Snakes
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image processing
scikit-image>=0.18.0
opencv-python>=4.5.0
aicsimageio>=4.0.0

# Plotting
matplotlib>=3.4.0

# Configuration management
PyYAML>=5.4.0

# Optional: For better performance
# numba>=0.54.0
